+++ PayPalNativeCheckout/src/test/java/com/braintreepayments/api/PayPalNativeCheckoutInternalClientUnitTest.java
@@ -37,7 +37,7 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
     private ClientToken clientToken;
     private TokenizationKey tokenizationKey;
 
-    private PayPalDataCollector payPalDataCollector;
+    private DataCollector dataCollector;
     private ApiClient apiClient;
 
     PayPalNativeCheckoutInternalClientCallback payPalInternalClientCallback;
@@ -49,7 +49,7 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
         clientToken = mock(ClientToken.class);
         tokenizationKey = mock(TokenizationKey.class);
 
-        payPalDataCollector = mock(PayPalDataCollector.class);
+        dataCollector = mock(DataCollector.class);
         apiClient = mock(ApiClient.class);
         payPalInternalClientCallback = mock(PayPalNativeCheckoutInternalClientCallback.class);
     }
@@ -63,9 +63,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .build();
         when(clientToken.getBearer()).thenReturn("client-token-bearer");
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutVaultRequest payPalRequest = new PayPalNativeCheckoutVaultRequest();
         payPalRequest.setBillingAgreementDescription("Billing Agreement Description");
@@ -108,9 +107,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .build();
         when(clientToken.getBearer()).thenReturn("client-token-bearer");
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutLineItem item =
                 new PayPalNativeCheckoutLineItem(PayPalNativeCheckoutLineItem.KIND_DEBIT, "Item 0",
@@ -181,9 +179,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .build();
         when(tokenizationKey.getBearer()).thenReturn("tokenization-key-bearer");
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutVaultRequest payPalRequest = new PayPalNativeCheckoutVaultRequest();
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
@@ -207,9 +204,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .authorizationSuccess(tokenizationKey)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutVaultRequest payPalRequest = new PayPalNativeCheckoutVaultRequest();
         payPalRequest.setDisplayName("");
@@ -233,9 +229,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .authorizationSuccess(tokenizationKey)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutVaultRequest payPalRequest = new PayPalNativeCheckoutVaultRequest();
         payPalRequest.setLocaleCode(null);
@@ -259,9 +254,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .authorizationSuccess(tokenizationKey)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutVaultRequest payPalRequest = new PayPalNativeCheckoutVaultRequest();
         payPalRequest.setMerchantAccountId(null);
@@ -285,9 +279,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .authorizationSuccess(tokenizationKey)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutVaultRequest payPalRequest = new PayPalNativeCheckoutVaultRequest();
         payPalRequest.setBillingAgreementDescription("");
@@ -311,9 +304,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .authorizationSuccess(tokenizationKey)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutRequest payPalRequest = new PayPalNativeCheckoutRequest("1.00");
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
@@ -335,9 +327,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .authorizationSuccess(tokenizationKey)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutRequest payPalRequest = new PayPalNativeCheckoutRequest("1.00");
         payPalRequest.setLineItems(new ArrayList<PayPalNativeCheckoutLineItem>());
@@ -354,10 +345,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
     }
 
     @Test
-    public void sendRequest_whenRiskCorrelationIdNotNull_setsClientMetadataIdToRiskCorrelationId()
-            throws Exception {
-        when(payPalDataCollector.getClientMetadataId(context, configuration)).thenReturn(
-                "sample-client-metadata-id");
+    public void sendRequest_whenRiskCorrelationIdNotNull_setsClientMetadataIdToRiskCorrelationId() throws Exception {
+        when(dataCollector.getClientMetadataId(context, configuration)).thenReturn("sample-client-metadata-id");
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
                 .configuration(Configuration.fromJson(Fixtures.CONFIGURATION_WITH_LIVE_PAYPAL))
@@ -365,9 +354,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .sendPOSTSuccessfulResponse(Fixtures.PAYPAL_HERMES_RESPONSE)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutRequest payPalRequest = new PayPalNativeCheckoutRequest("1.00");
         payPalRequest.setRiskCorrelationId("risk-correlation-id");
@@ -383,10 +371,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
     }
 
     @Test
-    public void sendRequest_whenRiskCorrelationIdNull_setsClientMetadataIdFromPayPalDataCollector()
-            throws Exception {
-        when(payPalDataCollector.getClientMetadataId(any(), any())).thenReturn(
-                "sample-client-metadata-id");
+    public void sendRequest_whenRiskCorrelationIdNull_setsClientMetadataIdFromPayPalDataCollector() throws Exception {
+        when(dataCollector.getClientMetadataId(any(), any())).thenReturn("sample-client-metadata-id");
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
                 .configuration(Configuration.fromJson(Fixtures.CONFIGURATION_WITH_LIVE_PAYPAL))
@@ -394,9 +380,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .sendPOSTSuccessfulResponse(Fixtures.PAYPAL_HERMES_RESPONSE)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutRequest payPalRequest = new PayPalNativeCheckoutRequest("1.00");
 
@@ -419,9 +404,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .authorizationSuccess(tokenizationKey)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutRequest payPalRequest = new PayPalNativeCheckoutRequest("1.00");
         payPalRequest.setShouldRequestBillingAgreement(false);
@@ -440,10 +424,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
     }
 
     @Test
-    public void sendRequest_withPayPalVaultRequest_callsBackPayPalResponseOnSuccess()
-            throws Exception {
-        when(payPalDataCollector.getClientMetadataId(any(), any())).thenReturn(
-                "sample-client-metadata-id");
+    public void sendRequest_withPayPalVaultRequest_callsBackPayPalResponseOnSuccess() throws Exception {
+        when(dataCollector.getClientMetadataId(any(), any())).thenReturn("sample-client-metadata-id");
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
                 .configuration(Configuration.fromJson(Fixtures.CONFIGURATION_WITH_LIVE_PAYPAL))
@@ -452,9 +434,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .sendPOSTSuccessfulResponse(Fixtures.PAYPAL_HERMES_BILLING_AGREEMENT_RESPONSE)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutVaultRequest payPalRequest = new PayPalNativeCheckoutVaultRequest();
         payPalRequest.setMerchantAccountId("sample-merchant-account-id");
@@ -476,10 +457,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
     }
 
     @Test
-    public void sendRequest_withPayPalCheckoutRequest_callsBackPayPalResponseOnSuccess()
-            throws Exception {
-        when(payPalDataCollector.getClientMetadataId(any(), any())).thenReturn(
-                "sample-client-metadata-id");
+    public void sendRequest_withPayPalCheckoutRequest_callsBackPayPalResponseOnSuccess() throws Exception {
+        when(dataCollector.getClientMetadataId(any(), any())).thenReturn("sample-client-metadata-id");
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
                 .configuration(Configuration.fromJson(Fixtures.CONFIGURATION_WITH_LIVE_PAYPAL))
@@ -488,9 +467,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .returnUrlScheme("sample-scheme")
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutRequest payPalRequest = new PayPalNativeCheckoutRequest("1.00");
         payPalRequest.setIntent("authorize");
@@ -520,9 +498,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .sendPOSTSuccessfulResponse(Fixtures.PAYPAL_HERMES_RESPONSE)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutRequest payPalRequest = new PayPalNativeCheckoutRequest("1.00");
         payPalRequest.setIntent("authorize");
@@ -544,9 +521,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .sendPOSTSuccessfulResponse(Fixtures.PAYPAL_HERMES_BILLING_AGREEMENT_RESPONSE)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutVaultRequest payPalRequest = new PayPalNativeCheckoutVaultRequest();
 
@@ -566,9 +542,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .sendPOSTErrorResponse(httpError)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutRequest payPalRequest = new PayPalNativeCheckoutRequest("1.00");
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
@@ -584,9 +559,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .sendPOSTSuccessfulResponse("{bad:")
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutRequest payPalRequest = new PayPalNativeCheckoutRequest("1.00");
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
@@ -601,9 +575,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .authorizationError(authError)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutRequest payPalRequest = new PayPalNativeCheckoutRequest("1.00");
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
@@ -619,9 +592,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
                 .configurationError(configurationError)
                 .build();
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         PayPalNativeCheckoutRequest payPalRequest = new PayPalNativeCheckoutRequest("1.00");
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
@@ -636,9 +608,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
         PayPalNativeCheckoutResultCallback callback =
                 mock(PayPalNativeCheckoutResultCallback.class);
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         sut.tokenize(payPalAccount, callback);
 
@@ -656,9 +627,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
         PayPalNativeCheckoutResultCallback callback =
                 mock(PayPalNativeCheckoutResultCallback.class);
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         sut.tokenize(payPalAccount, callback);
 
@@ -683,9 +653,8 @@ public class PayPalNativeCheckoutInternalClientUnitTest {
         PayPalNativeCheckoutResultCallback callback =
                 mock(PayPalNativeCheckoutResultCallback.class);
 
-        PayPalNativeCheckoutInternalClient sut =
-                new PayPalNativeCheckoutInternalClient(braintreeClient, payPalDataCollector,
-                        apiClient);
+        PayPalNativeCheckoutInternalClient sut = new PayPalNativeCheckoutInternalClient(braintreeClient,
+                dataCollector, apiClient);
 
         sut.tokenize(payPalAccount, callback);