+++ CHANGELOG.md
@@ -9,6 +9,11 @@
   * UnionPay
     * Remove `union-pay` module
       * UnionPay cards can now be processed as regular cards (through the `card` module) due to their partnership with Discover
+  * BraintreeDataCollector
+    * Replace `DataCollector#collectDeviceData(context, merchantId, callback)` with 
+      `DataCollector#collectDeviceData(context, riskCorrelationId, callback)`
+  * PayPalDataCollector
+    * Remove `paypal-data-collector` module (use `data-collector`)
   * Venmo
     * Remove `VenmoListener`, `VenmoTokenizeAccountCallback`
     * Add `VenmoLauncher`, `VenmoAuthChallenge`, `VenmoAuthChallengeCallback`,